Falschen Argument-Typ für IMatrix#setField() gefixt

+++ UnifiedVectors/src/main/java/de/m_marvin/unimat/api/IMatrix.java
@@ -5,7 +5,7 @@ public interface IMatrix<M extends IMatrix<?>> {
 	public M copy();
 	
 	public float getField(int x, int y);
-	public void setField(int x, int y, int f);
+	public void setField(int x, int y, float f);
 
 	public float[] toFloatArr();
 	public void loadFloatArr(float[] arr);
